Feature/comment

+++ src/main/java/com/example/week06/domain/community/service/CommentService.java
@@ -5,49 +5,48 @@ import com.example.week06.domain.community.entity.Post;
 import com.example.week06.domain.user.entity.User;
 import com.example.week06.domain.community.repository.CommentRepository;
 import com.example.week06.domain.community.repository.PostRepository;
-import com.example.week06.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-
-import javax.transaction.Transactional;
-
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 @RequiredArgsConstructor
 public class CommentService {
     private final CommentRepository commentRepository;
     private final PostRepository postRepository;
-    private final UserRepository userRepository;
 
-    @Transactional
-    public void createComment(CommentRequest commentRequest, String email, Long postId){
-        User user = userRepository.findByEmail(email).get();
-        Post post = postRepository.findById(postId).get();
-        Comment comment = new Comment(commentRequest, user, post);
+    public void createComment(String content, Long postId, User user){
+        Post post = postRepository.findById(postId).orElseThrow(
+                ()-> new ResponseStatusException(HttpStatus.NOT_FOUND,"존재하지 않는 게시물입니다.")
+        );
+        Comment comment = new Comment(content, user, post);
+
         commentRepository.save(comment);
     }
 
-    @Transactional
-    public void deleteComment(Long commentId, String email){
-        User user = userRepository.findByEmail(email).orElseThrow(
-                () -> new IllegalArgumentException("사용자가 존재하지 않습니다.")
+    public void deleteComment(Long commentId, User user) {
+        Comment comment = commentRepository.findById(commentId).orElseThrow(
+                ()-> new ResponseStatusException(HttpStatus.NOT_FOUND,"댓글이 존재하지 않습니다.")
         );
 
-        commentRepository.findByIdAndUser(commentId, user).orElseThrow(
-                () -> new IllegalArgumentException("댓글이 존재하지 않습니다.")
-        );
+        if(!comment.getUser().getId().equals(user.getId())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "댓글에 대한 권한이 없습니다.");
+        }
 
         commentRepository.deleteById(commentId);
     }
 
-    @Transactional
-    public void putComment(Long commentId, CommentRequest commentRequest, String email){
+    public void putComment(Long commentId, String content, User user) {
         Comment comment = commentRepository.findById(commentId).orElseThrow(
-                ()-> new IllegalArgumentException("댓글이 존재하지 않습니다."));
+                ()-> new ResponseStatusException(HttpStatus.NOT_FOUND,"댓글이 존재하지 않습니다."));
 
-        User writer = comment.getUser();
+        if(!comment.getUser().getId().equals(user.getId())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "댓글에 대한 권한이 없습니다.");
+        }
 
-
+        comment.update(content);
+        commentRepository.save(comment);
     }
 
 }
